Moved createAssignmentQueueItems into assignment queue service
